requirements.txt
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
coverage==7.13.4
cryptography==46.0.5
google-auth==2.48.0
google-auth-oauthlib==1.3.0
greenlet==3.3.2
gspread==6.2.1
gspread-dataframe==4.0.0
httplib2==0.31.2
idna==3.11
iniconfig==2.3.0
numpy==2.4.2
oauth2client==4.1.3
oauthlib==3.3.1
packaging==26.0
pandas==3.0.1
pluggy==1.6.0
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.48
typing_extensions==4.15.0
urllib3==2.6.3
